Bugfix in general record editor
when a new field was added, it didn't show when exiting the editing mode.
New feature: double-click a record in a table, to enter a limited ("in place") editor.
Corresponding changes to page_viewer.htm, incl. new JS function absorb_item_updates_REVISED(), and to Vue components 'vue-plugin-table' and 'vue-plugin-single-record'